Corrige estouro de variável

+++ Server/SoftwareParameters.cs
@@ -70,13 +70,13 @@ namespace TrueMiningDesktop.Server
         public decimal ShareCoef { get; set; } = 1;
 
         [JsonProperty("shareMmc", NullValueHandling = NullValueHandling.Ignore)]
-        public short ShareMmc { get; set; } = 1;
+        public decimal ShareMmc { get; set; } = 1;
 
         [JsonProperty("defaultHashMuString", NullValueHandling = NullValueHandling.Ignore)]
         public string DefaultHashMuString { get; set; } = "H/s";
 
         [JsonProperty("defaultHashMuCoef", NullValueHandling = NullValueHandling.Ignore)]
-        public short DefaultHashMuCoef { get; set; } = 1;
+        public int DefaultHashMuCoef { get; set; } = 1;
     }
 
     public partial class PaymentCoin
